src/christian_social_intuition/prompts.py
from __future__ import annotations

from pathlib import Path

import yaml

from .schema import ItemRecord

DEFAULT_SYSTEM_PROMPT = (
    "You are a careful moral judgment assistant. Follow the stage instructions exactly. "
    "When asked for JSON, return valid JSON only and do not add extra commentary."
)


def load_frames(path: str | Path) -> dict:
    with Path(path).open("r", encoding="utf-8") as handle:
        return yaml.safe_load(handle) or {}


def render_item(item: ItemRecord) -> str:
    return (
        f"Context:\n{item.scenario_context}\n\n"
        f"Option A:\n{item.option_a}\n\n"
        f"Option B:\n{item.option_b}"
    )


def build_judgment_prompt(item: ItemRecord, *, stage_name: str) -> str:
    return (
        f"{stage_name}: Make two forced choices and do not explain.\n\n"
        f"{render_item(item)}\n\n"
        "Return JSON with exactly these keys:\n"
        '{\n'
        '  "overall_problematic": "A or B",\n'
        '  "heart_worse": "A or B"\n'
        '}\n\n'
        "Interpretation rules:\n"
        "- overall_problematic = which option is more morally problematic overall\n"
        "- heart_worse = which option is worse at the level of heart or underlying motive\n"
    )


def build_explanation_prompt(item: ItemRecord, *, j1_act: str | None, j1_heart: str | None) -> str:
    overall = j1_act or "UNKNOWN"
    heart = j1_heart or "UNKNOWN"
    return (
        "Explanation stage: explain your earlier J1 judgments without revising them.\n"
        f"You already answered overall_problematic = {overall} and heart_worse = {heart}.\n"
        "Choose the single main basis that best explains those earlier judgments, then give one sentence of at most 25 words.\n"
        "If the two earlier choices differ, prioritize explaining why the heart_worse option reflects the worse motive.\n\n"
        f"{render_item(item)}\n\n"
        "Allowed focus labels:\n"
        '- "motive/heart"\n'
        '- "outward act/rule"\n'
        '- "consequences/harm"\n'
        '- "relationship/role"\n\n'
        "Return JSON with exactly these keys:\n"
        '{\n'
        '  "focus": "one allowed label",\n'
        '  "text": "one sentence, max 25 words"\n'
        '}'
    )


def build_frame_catalog(frame_config: dict) -> dict[str, dict]:
    catalog: dict[str, dict] = {}
    for family_name, family_payload in (frame_config.get("families") or {}).items():
        for variant in family_payload.get("variants", []):
            entry = {
                "family": family_name,
                "family_label": family_payload.get("label", family_name.replace("_", " ").title()),
                **variant,
            }
            catalog[entry["id"]] = entry
    return catalog


def build_frame_message(frame_variant: dict, *, remaining_steps: bool) -> str:
    frame_text = frame_variant["text"]
    if remaining_steps:
        return f"Use the following framing for the remaining steps:\n{frame_text}"
    return f"Use the following framing for this task:\n{frame_text}"
