pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "christian-social-intuition"
version = "0.1.0"
description = "Stage-separated Christian framing x Social Intuitionist experiment harness for LLMs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.8",
  "pandas>=2.1",
  "PyYAML>=6.0",
  "statsmodels>=0.14",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
csi-sim = "christian_social_intuition.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
